requirements.txt
torch=='1.7.0'
torchvision=='0.8.2'
numpy=='1.19.2'
sklearn=='0.24.0'
torchmeta=='1.6.1'
pyyaml=='5.3.1'
